Add support for organization_updated event in WorkOS webhook
