+++ front/lib/api/workspace_domains.ts
@@ -1,6 +1,6 @@
 import { Workspace } from "@app/lib/models/workspace";
 import { WorkspaceHasDomainModel } from "@app/lib/models/workspace_has_domain";
-import type { LightWorkspaceType, Result } from "@app/types";
+import type { LightWorkspaceType, Result, WorkspaceDomain } from "@app/types";
 import { Err, Ok } from "@app/types";
 
 export async function upsertWorkspaceDomain(
@@ -59,4 +59,24 @@ export async function deleteWorkspaceDomain(
   await existingDomain.destroy();
 
   return new Ok(undefined);
+}
+
+export async function getWorkspaceVerifiedDomains(
+  workspace: LightWorkspaceType
+): Promise<WorkspaceDomain[]> {
+  const workspaceDomains = await WorkspaceHasDomainModel.findAll({
+    attributes: ["domain", "domainAutoJoinEnabled"],
+    where: {
+      workspaceId: workspace.id,
+    },
+  });
+
+  if (workspaceDomains) {
+    return workspaceDomains.map((domain) => ({
+      domain: domain.domain,
+      domainAutoJoinEnabled: domain.domainAutoJoinEnabled,
+    }));
+  }
+
+  return [];
 }
