+++ front/pages/api/v1/w/[wId]/verified_domains.ts
@@ -2,10 +2,10 @@ import type { GetWorkspaceVerifiedDomainsResponseType } from "@dust-tt/client";
 import type { NextApiRequest, NextApiResponse } from "next";
 
 import { withPublicAPIAuthentication } from "@app/lib/api/auth_wrappers";
-import { getWorkspaceVerifiedDomains } from "@app/lib/api/workspace";
 import type { Authenticator } from "@app/lib/auth";
 import { apiError } from "@app/logger/withlogging";
 import type { WithAPIErrorResponse } from "@app/types";
+import { getWorkspaceVerifiedDomains } from "@app/lib/api/workspace_domains";
 
 /**
  * @ignoreswagger
