+++ front/pages/poke/[wId]/index.tsx
@@ -30,10 +30,7 @@ import { ActiveSubscriptionTable } from "@app/components/poke/subscriptions/tabl
 import { TrackerDataTable } from "@app/components/poke/trackers/table";
 import { WorkspaceInfoTable } from "@app/components/poke/workspace/table";
 import config from "@app/lib/api/config";
-import {
-  getWorkspaceCreationDate,
-  getWorkspaceVerifiedDomains,
-} from "@app/lib/api/workspace";
+import { getWorkspaceCreationDate } from "@app/lib/api/workspace";
 import { useSubmitFunction } from "@app/lib/client/utils";
 import { withSuperUserAuthRequirements } from "@app/lib/iam/session";
 import { Plan, Subscription } from "@app/lib/models/plan";
@@ -50,6 +47,7 @@ import type {
   WorkspaceType,
 } from "@app/types";
 import { WHITELISTABLE_FEATURES } from "@app/types";
+import { getWorkspaceVerifiedDomains } from "@app/lib/api/workspace_domains";
 
 export const getServerSideProps = withSuperUserAuthRequirements<{
   activeSubscription: SubscriptionType;
